requirements.txt
bleak
opencv-python
customtkinter
pillow